(Add) : throw exceptions for User, Auth and Category

+++ event-manager-backend/src/main/java/com/dauphine/event_manager_backend/service/impl/EventServiceImpl.java
@@ -1,7 +1,9 @@
 package com.dauphine.event_manager_backend.service.impl;
 
 import com.dauphine.event_manager_backend.exceptions.CategoryNotFoundByIdException;
+import com.dauphine.event_manager_backend.exceptions.EventNameAlreadyExistsException;
 import com.dauphine.event_manager_backend.exceptions.EventNotFoundByIdException;
+import com.dauphine.event_manager_backend.exceptions.UserNotFoundByIdException;
 import com.dauphine.event_manager_backend.model.Category;
 import com.dauphine.event_manager_backend.model.Event;
 import com.dauphine.event_manager_backend.model.User;
@@ -9,11 +11,12 @@ import com.dauphine.event_manager_backend.repository.CategoryRepository;
 import com.dauphine.event_manager_backend.repository.EventRepository;
 import com.dauphine.event_manager_backend.repository.UserRepository;
 import com.dauphine.event_manager_backend.service.EventService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
@@ -45,33 +48,33 @@ public class EventServiceImpl implements EventService {
     }
 
     @Override
-    public Event create(String title, String city, String address , LocalDateTime date, String description, UUID categoryId, UUID userId) {
-        Optional<Category> optionalCategory = categoryRepository.findById(categoryId);
-        Optional<User> optionalUser =userRepository.findById(userId);
-        if (optionalCategory.isPresent() && optionalUser.isPresent()) {
-            Category category = optionalCategory.get();
-            User user = optionalUser.get();
-            Event newEvent = new Event(title, city, address, date, description, category, user);
-            return eventRepository.save(newEvent);
+    public Event create(String title, String city, String address , LocalDateTime date, String description, UUID categoryId, UUID userId) throws CategoryNotFoundByIdException, UserNotFoundByIdException, EventNameAlreadyExistsException {
+        Category category = categoryRepository.findById(categoryId)
+                    .orElseThrow(() -> new CategoryNotFoundByIdException(categoryId));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundByIdException(userId));
+        if (eventRepository.existsByName(title)) {
+            throw new EventNameAlreadyExistsException(title);
         }
-        else if (optionalCategory.isEmpty()) {
-            throw new RuntimeException("Category with following id not found : " + categoryId);
-        }
-        else {
-            throw new RuntimeException("User with following id not found : " + userId);
-        }
+        Event newEvent = new Event(title, city, address, date, description, category, user);
+        return eventRepository.save(newEvent);
     }
 
     @Override
     public void deleteById(UUID id) throws EventNotFoundByIdException {
-        getEventById(id);
+        if (!eventRepository.existsById(id)) {
+            throw new EventNotFoundByIdException(id);
+        }
         eventRepository.deleteById(id);
     }
 
     @Override
-    public Event update(UUID eventId, String title, String city, String address ,LocalDateTime date, String description, UUID categoryId, UUID userId) throws EventNotFoundByIdException, CategoryNotFoundByIdException {
+    public Event update(UUID eventId, String title, String city, String address ,LocalDateTime date, String description, UUID categoryId, UUID userId) throws EventNotFoundByIdException, CategoryNotFoundByIdException, EventNameAlreadyExistsException {
         Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new CategoryNotFoundByIdException(categoryId));
         Event event = getEventById(eventId);
+        if (eventRepository.existsByName(title) && !Objects.equals(event.getTitle(), title)) {
+            throw new EventNameAlreadyExistsException(title);
+        }
         event.setTitle(title);
         event.setCity(city);
         event.setAddress(address);
